refactor: update enum imports to value_objects
